FEAT-#2197: move reshape.py content into general.py
